+++ bin/build.sh
@@ -2,8 +2,6 @@
 
 yarn typecheck
 yarn lint
-# yarn test
 
 rm -rf dist
 npm run tsc -- -p .
-rm -fr dist/__tests__