+++ CHANGELOG.md
@@ -18,3 +18,6 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 - Initial safe SDKs support. We support all actions and fetch state.
 - Orderbook support for trading of generic collections.
 - Orderbook events support
+- Option to sort orderbook bids and asks by price.
+- Fetching and parsing `TradeIntermediary` object by its ID.
+- Returning list of `TradePayment` IDs when finishing a trade.